Fix MCMCGetCurrentChain with multithreading
In a parallel openMP section, we look up the current chain based on
thread. The map has to be managed by us. This allows bat to run inside
of another openMP parallel section such that thread indices may exceed
the number of chains. To unify the handling, I removed
`BCEngineMCMC::MCMCGetThreadNum()` and
`BCEngineMCMC::fMCMCCurrentChain`. Whenever the chain index is needed
now, it has to come from `BCEngineMCMC::MCMCGetCurrentChain` and the
thread id should be irrelevant.

The map is updated by `BCEngineMCMC::UpdateChainIndex` and has to be
cleared at appropriate places. I'd really prefer this method to be
private but the `BCPriorModel` needs to call it. I found no better
solution than making it public and adding a warning in the doxygen
comment.

If no threading is used (no mcmc or just openmp disabled), the current
chain is 0. This should make it easy to write code agnostic of the
parallel features. Before `MCMCGetCurrentChain` returned -1, which was
never actually used or checked for anywhere and caused the return value
to be an `int`, now it's `unsigned`.

Some refactoring is included where I stumbled upon it. I removed all
warnings appearing with `--enable-debug` and successfully ran
`OMP_NUM_THREADS=2 make check`